requirements.txt
absl-py==2.0.0
accelerate==0.21.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.6.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.40.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
datasets==2.16.1
deepspeed==0.12.6
dgl==2.3.0+cu121
dill==0.3.7
distro==1.9.0
einops==0.7.0
exceptiongroup==1.2.2
filelock==3.13.1
fire==0.5.0
flash-attn==2.4.2
frozenlist==1.4.1
fsspec==2023.10.0
gin-config==0.5.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
grpcio==1.60.0
h11==0.14.0
hf-xet==1.1.1
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.31.1
idna==3.6
Jinja2==3.1.2
joblib==1.4.2
Markdown==3.5.1
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.9.0
packaging==23.2
pandas==2.1.4
peft==0.7.1
pillow==11.2.1
protobuf==4.23.4
psutil==5.9.7
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pynvml==11.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.5.1
scipy==1.11.4
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.1
some-package==0.1
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.15.2
torch==2.1.2
torchdata==0.7.1
tqdm==4.66.1
transformers==4.37.0
triton==2.1.0
trl==0.4.7
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
